Move most of the fiber logic into a cpp file
To make fiber.h a bit leaner to include and easier to read
