+++ .gitignore
@@ -9,6 +9,7 @@ cmake-build*/
 
 # Visual Studio local files
 .vs/
+Folder.DotSettings.user
 
 # Lint output files
 clang-tidy-output.txt
